fixed duplication and improve styles
